+++ src/client-types.ts
@@ -11,6 +11,7 @@ import type { DeviceKind, DeviceTarget, Platform, PlatformSelector } from './uti
 import type { FindLocator } from './utils/finders.ts';
 import type { ScreenshotOverlayRef, SnapshotNode, SnapshotVisibility } from './utils/snapshot.ts';
 import type { MetroPrepareKind, PrepareMetroRuntimeResult } from './client-metro.ts';
+import type { MetroBridgeScope } from './client-metro-companion-contract.ts';
 
 export type { FindLocator } from './utils/finders.ts';
 
@@ -270,11 +271,7 @@ export type MetroPrepareOptions = {
   publicBaseUrl: string;
   proxyBaseUrl?: string;
   bearerToken?: string;
-  bridgeScope?: {
-    tenantId: string;
-    runId: string;
-    leaseId: string;
-  };
+  bridgeScope?: MetroBridgeScope;
   launchUrl?: string;
   companionProfileKey?: string;
   companionConsumerKey?: string;
