--- src/platforms/linux/index.ts
@@ -1,1 +1,0 @@
-export { snapshotLinux } from './snapshot.ts';
